Create edges _before_ vertices when generating
It turns out that when determining how many nodes to generate in a
relationship, you need to account for the number of _existing_
relationships that are already pointing to or from a vertex. An
example of this is a an Author that has many articles, but where each
article can have many authors.

So your generation needs to proceed as

1. start generating the first author
2. start generating each article for that first author
3. start generating the other authors for that article

But let's say it is determined that the article is going to have three
authors, then we need to know that the first author has already been
created, and that we only need to generate two more.

For reason, when generating a path to a related vertex, we actually
need to create the edge _first_. That way, the generation of the
vertex can know that at least one part of the relationship is
satisfied.

To accomplish this, we allow for the creating process to allocate the
`id` of the vertex, and pass it in, rather than having the vertex
creation allocate its own id