Set default affinity to zero.
Having affinity is important for circular relationships, but for
directed relationships you almost always want to have a unique
target. For example, if you have a person that has a github account,
then you want their github account to be unique. The introduction of
the concept of graph affinity also saw the introduction of a default
affinity for _all_ relationships. This is counter-intuitive, and
clearly backwards compatible because it breaks the behavior of current
generations by making formerly unique relationships shared between vertices.

This makes the default affinity 0, and as a result circular
relationships can still blow up and must be clearly annotated with a
value for affinity.