Allow generating multiple relationships from a vertex
There is really nasty bug where a vertex counts the number of
_existing_ edges towards the size limit of how many it wants to
generate. So, for example, if we have a user with two friends, and
then we generate a set of articles that this person has written (Let's
say that graphgen decides it should be three articles). It will count
the two friend relationships towards the article total, and instead of
generating three articles, it will only generate one. This is
especially pernicious with `constant` relationships of one, since if
there is any other edge connecting that vertex, it will not be
generated.

The answer is only use existing edges _of the same edge type_ towards
the total number of edges that already exist.