Convert to Deno
Deno is awesome, and is getting better by the day. The development
cycle is extraordinarily fast, and because it can emit an npm
package from a deno module, there is no need to use node as the
primary javacript runtime for development when we have a pure JS
library like this. It gets trickier when you have runtime specific
apis on which you depend, but since this has none, we're good.

This uses the same template laid out in
https://github.com/thefrontside/continuation