Add `__typename` properties to `Node` API.
When simulating the backstage catalog API, you need to have the `kind`
parameter in the entity body which is the same as the type of the
graphql gen.

graphql:

```graphql
type Component {
 name: String!
 description: String!
}

backstage entity:

```json
{
  "kind": "Component",
  "metadata": {
    "name": "artists-lookup",
    "namespace": "default"
  },
  spec: {
    "description": "Artists Lookup Service"
  }
}
```

But currently the string representing the graphql type name is not
available to use.

This adds a `__typename` property which is prefixed to indicate that
it is mostly for metaprogramming. This property is deliberately not
enumerable so that it doesn't make debug output unsightly, and also so
that you can iterate over just the "concrete" properties without
making exceptions for the `__typename`