Implement a basic CLI build
We want a way to be able to visually represent a graphgen factory so
that you can see what kinds of data it will generate for a given set
of inputs.

We can distribute the tool as a CLI compiled for each individual
platform using `deno compile`.

This creates a CLI build for each platform and compiles it with only
the ability to access the network, and also to allow reading from the
filesystem. Binaries are created for each platform, and uploaded as a
github release on any tag starting with `graphgen-cli-vx.x.x`