Handle default commonjs exports
CJS has both the case where the export object *is* the default, and
also where there is a module export called "default" and we have to
handle them both.

This just checks and if there is a property called `default` we use
that instead. Again, we need to have some form of validation on the
export itself but this will be flexible for now.