Make start command run local development
With the integration to our website, you need to use a local netlify
development environment to do local dev. This is really cool in that
it simulates how the app will run in production, but you have to know
to run `npx netlify dev`

This makes that command the default `start` script, so that `npm
start` goes through netlify.