Add docs for basic fields.
Now that we have a site in place, we need to actually place the docs
there.

This starts with describing fields, how to generate them, how to
register custom generators.

I've also left place-holders for the next pieces of content to write,
but have not wired them up yet so that there are "no broken windows"