⬆️Upgrade `graphql` and `gaussian` dependencies
The [test were failing when updating the website][1] because we had to
use some specifics of the way `https://esm.sh` bundled graphql-js by
reaching into the `/graphql` directory of the npm package directly.

However, now that [GraphQL-js officially supports Deno][2] we can
upgrade it. This also resolves a dependency in the gaussian
distribution library we used to pin it at a single version.

[1]: https://github.com/thefrontside/graphgen/actions/runs/4876196677/jobs/8699372244
[2]: https://github.com/graphql/graphql-js/pull/3741