Fix bug in base redirects
The current redirect to the graphgen site is broken when coming
through the new frontside.com router.

It turns out that the redirect location when going from `/` ->
`docs/introduction` did not properly incorporate the base url when
relocating, So going to
https://frontside.com/graphgen/docs/introduction directly worked, but
going to
https://frontside.com/graphgen did not.

This makes sure in the index route that the redirect properly
incorporates the base path into the redirected location, and then for
good measure, makes sure base links are ok whether they end in `/` or
not.